initial move of db folder and schema refactor
