+++ backend/src/database/pyproject.toml
@@ -2,7 +2,7 @@
 name = "database"
 version = "0.1.0"
 description = ""
-authors = ["Noah Hayes <nh869@bath.ac.uk>"]
+authors = []
 license = "MIT"
 readme = "README.md"
 packages = [
@@ -11,11 +11,7 @@ packages = [
 
 [tool.poetry.dependencies]
 python = ">=3.13"
-sqlalchemy = ">=2.0.39,<3.0.0"
 
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.3.5"
-pytest-sqlalchemy = "^0.2.1"
 
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
