+++ backend/src/database/scripts/get_graph_data.py
